req changes
