--- src/common/__utils/constants.ts
@@ -1,8 +1,0 @@
-export const CHARACTER_LIMIT_50 = 50;
-export const CHARACTER_LIMIT_49 = 49;
-export const CHARACTER_LIMIT_48 = 48;
-export const THERE_IS_A_PROBLEM_ERROR: string = "There is a problem";
-export const USER_DATA = "user";
-export const GET_ACSP_REGISTRATION_DETAILS_ERROR: string = "Unable to get registration details for the transaction";
-export const POST_ACSP_REGISTRATION_DETAILS_ERROR: string = "Unable to post registration details for the transaction";
-export const PREVIOUS_PAGE_URL: string = "previouspageurl";
