+++ src/utils/constants.ts
@@ -1,0 +1,2 @@
+export const USER_DATA = "user";
+export const PREVIOUS_PAGE_URL: string = "previouspageurl";
