+++ src/utils/sessionHelper.ts
